move components into their own files
