+++ src/frontend/src/views/Data/components/SamplesView/components/SamplesTable/components/SortableHeader/components/TooltipText/index.tsx
@@ -1,0 +1,28 @@
+import { NewTabLink } from "src/common/components/library/NewTabLink";
+
+export type TooltipTextType = {
+  boldText: string;
+  regularText: string;
+  link: {
+    href: string;
+    linkText: string;
+  };
+};
+
+interface TooltipTextProps {
+  tooltipStrings: TooltipTextType;
+}
+
+export const TooltipText = ({
+  tooltipStrings,
+}: TooltipTextProps): JSX.Element => {
+  const { boldText, regularText, link } = tooltipStrings;
+  const { href, linkText } = link ?? {};
+
+  return (
+    <div>
+      <b>{boldText}</b>: {regularText}{" "}
+      {link && <NewTabLink href={href}>{linkText}</NewTabLink>}
+    </div>
+  );
+};
