+++ src/frontend/src/views/Data/components/SamplesView/components/SamplesTable/index.tsx
@@ -4,13 +4,11 @@ import {
   getCoreRowModel,
   getSortedRowModel,
   Getter,
-  Header,
   RowSelectionState,
   useReactTable,
 } from "@tanstack/react-table";
 import {
   CellComponent,
-  CellHeader,
   Icon,
   InputCheckbox,
   Table,
@@ -18,57 +16,19 @@ import {
   TableRow,
 } from "czifui";
 import { map } from "lodash";
-import { ReactNode, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { NewTabLink } from "src/common/components/library/NewTabLink";
 import { IdMap } from "src/common/utils/dataTransforms";
 import { datetimeWithTzToLocalDate } from "src/common/utils/timeUtils";
 import { LineageTooltip } from "../../../LineageTooltip";
+import { SortableHeader } from "./components/SortableHeader";
 import { StyledCellBasic, StyledPrivateId } from "./style";
 
-// TODO-TR (mlila): types
 interface Props {
   data: IdMap<Sample> | undefined;
   isLoading: boolean;
   setCheckedSamples(samples: Sample[]): void;
 }
-
-// TODO-TR (mlila): move this header component into its own file
-interface SortableProps {
-  header: Header<any, any>;
-  children: ReactNode & string;
-  tooltip?: any;
-}
-
-export const SortableHeader = ({
-  header,
-  children,
-  tooltip,
-  ...props
-}: SortableProps): JSX.Element => {
-  const { getCanSort, getIsSorted, getToggleSortingHandler } = header.column;
-
-  const sortable = getCanSort();
-  const sortDirection = getIsSorted() || undefined;
-  const handler = getToggleSortingHandler();
-
-  return (
-    <CellHeader
-      onClick={handler}
-      direction={sortDirection}
-      active={Boolean(sortDirection)}
-      hideSortIcon={!sortable}
-      shouldShowTooltipOnHover={Boolean(tooltip)}
-      tooltipProps={{
-        arrow: false,
-        sdsStyle: "light",
-        title: <TooltipText tooltip={tooltip} />,
-      }}
-      {...props}
-    >
-      {children}
-    </CellHeader>
-  );
-};
 
 // TODO-TR (mlila): move this default cell into its own component file
 const DefaultCell = ({ getValue }: { getValue: Getter<any> }): JSX.Element => (
@@ -79,19 +39,6 @@ const DefaultCell = ({ getValue }: { getValue: Getter<any> }): JSX.Element => (
     shouldShowTooltipOnHover={false}
   />
 );
-
-// TODO-TR (mlila): move this into its own component file
-const TooltipText = ({ tooltip }) => {
-  const { boldText, regularText, link } = tooltip;
-  const { href, linkText } = link ?? {};
-
-  return (
-    <div>
-      <b>{boldText}</b>: {regularText}{" "}
-      {link && <NewTabLink href={href}>{linkText}</NewTabLink>}
-    </div>
-  );
-};
 
 const columns: ColumnDef<Sample, any>[] = [
   {
@@ -140,7 +87,7 @@ const columns: ColumnDef<Sample, any>[] = [
       <SortableHeader
         header={header}
         shouldShowTooltipOnHover
-        tooltip={{
+        tooltipStrings={{
           boldText: "Private ID",
           regularText:
             "User-provided private ID. Only users in your Group can see it.",
@@ -181,7 +128,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "Public ID",
           regularText:
             "This is your GISAID ID or public ID generated by CZ Gen Epi.",
@@ -211,7 +158,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "Upload Date",
           regularText: "Date on which the sample was uploaded to CZ Gen Epi.",
         }}
@@ -234,7 +181,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "Collection Date",
           regularText:
             "User-provided date on which the sample was collected from an individual or an environment.",
@@ -252,7 +199,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "Lineage",
           link: {
             href: "https://cov-lineages.org/pangolin.html",
@@ -290,7 +237,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "Collection Location",
           regularText:
             "User-provided geographic location where the sample was collected (at the county level or above).",
@@ -315,7 +262,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "Sequencing Date",
           regularText: "User-provided date on which the sample was sequenced.",
         }}
@@ -331,7 +278,7 @@ const columns: ColumnDef<Sample, any>[] = [
     header: ({ header }) => (
       <SortableHeader
         header={header}
-        tooltip={{
+        tooltipStrings={{
           boldText: "GISAID Status",
           regularText:
             "Whether your sample has been Not Yet Submitted, Submitted, Accepted (with GISAID accession), Rejected, or Not Eligible (marked private).",
